Create Guessing Number Game
This is a basic python projects which uses basic concepts of python programming without using any python library. The project creates a program which could generate a random number by itself in the provided range of numbers. The user have to guess the number which generated by the random function and if the user enters wrong number it also gives hint to the user whether it is smaller or greater than the generated number.
Features:
 - Generates random number
 - Gives attempts until user enters the correct number
 - Gives hints to the user whether the entered number is far or nearer to the number generated
 - Uses basic concepts of python without using/importing any library
 - Easy to code and understand
 - Code is provided along with comments which helps others to understand easily
 - Proper Indentation